Added NATIVE binding support
